Refactored Form Widget
